apps/model-serving-api/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
pymilvus==2.3.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.26.1
